Create README.md
This project shows all the information about the weather of a specified city and country. The data obtained comes from queries to the openweather API. 
This application is developed in javascript and uses tailwind to establish the different styles with which the information is presented.
In this first version it is only possible to enter cities that are found only in the 9 countries added in the respective field. It is planned for future versions to take into account the full range of countries and develop a search that adjusts to the data entered by the user.
The main objective of this project is the consumption of information from an API. Showing the results of the queries in a pleasant way for the user to see.
It is worth mentioning that the application is responsive and condenses weather information that is of interest to users.